+++ src/Cache/Subscription/PurgeSubscriptionProvider.php
@@ -5,7 +5,6 @@ declare(strict_types=1);
 namespace Sofascore\PurgatoryBundle\Cache\Subscription;
 
 use Doctrine\Persistence\ManagerRegistry;
-use Opis\Closure\ReflectionClosure;
 use Psr\Container\ContainerInterface;
 use Sofascore\PurgatoryBundle\Attribute\RouteParamValue\ExpressionValues;
 use Sofascore\PurgatoryBundle\Attribute\RouteParamValue\PropertyValues;
@@ -161,7 +160,15 @@ final class PurgeSubscriptionProvider implements PurgeSubscriptionProviderInterf
 
     private function validateIfClosure(\Closure $expression, string $routeName, string $entity): void
     {
-        $reflection = new ReflectionClosure($expression);
+        $reflection = new \ReflectionFunction($expression);
+
+        if (null !== $reflection->getClosureThis()) {
+            throw new InvalidIfClosureException($routeName, 'Closure must be static');
+        }
+
+        if ([] !== $reflection->getClosureUsedVariables()) {
+            throw new InvalidIfClosureException($routeName, 'Closure must not capture variables');
+        }
 
         $returnType = $reflection->getReturnType();